requirements.txt
pyspark==3.2.4
pymongo==4.3.3
flask==2.3.2
flask-cors==4.0.0
werkzeug==2.3.3
pandas==1.5.0
numpy==1.23.0
scikit-learn==1.1.2
requests==2.28.1
python-dotenv==0.21.0
joblib==1.2.0
apscheduler==3.10.1
gunicorn==20.1.0